Menu

+++ src/main/java/com/uplog/uplog/domain/post/application/PostService.java
@@ -1,5 +1,7 @@
 package com.uplog.uplog.domain.post.application;
 
+import com.uplog.uplog.domain.comment.dao.CommentRepository;
+import com.uplog.uplog.domain.like.dao.PostLikeRepository;
 import com.uplog.uplog.domain.member.dao.MemberRepository;
 import com.uplog.uplog.domain.member.model.Member;
 import com.uplog.uplog.domain.member.model.Position;
@@ -16,6 +18,7 @@ import com.uplog.uplog.domain.project.model.Project;
 import com.uplog.uplog.domain.task.exception.NotFoundTaskByIdException;
 import com.uplog.uplog.global.exception.AuthorityException;
 import com.uplog.uplog.global.exception.NotFoundIdException;
+import com.uplog.uplog.global.method.AuthorizedMethod;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -33,6 +36,9 @@ public class PostService {
     private final ProductRepository productRepository;
     private final ProjectRepository projectRepository;
     private final MenuRepository menuRepository;
+    private final AuthorizedMethod authorizedMethod;
+    private final PostLikeRepository postLikeRepository;
+    private final CommentRepository commentRepository;
 
     /*
     Create
@@ -42,6 +48,9 @@ public class PostService {
         Member author = memberRepository.findMemberById(id)
                 .orElseThrow(() -> new NotFoundIdException("해당 멤버는 존재하지 않습니다."));
 
+        //현재 프로젝트 팀 내에 존재하는 멤버,기업이 아닌 회원,클라이언트가 아닌 멤버 확인
+        authorizedMethod.CreatePostTaskValidateByMemberId(author);
+
         Menu menu = menuRepository.findById(createPostRequest.getMenuId())
                 .orElseThrow(() -> new NotFoundIdException("해당 메뉴는 존재하지 않습니다."));
 
@@ -50,6 +59,7 @@ public class PostService {
 
         Product product = productRepository.findById(createPostRequest.getProductId())
                 .orElseThrow(() -> new NotFoundIdException("해당 제품은 존재하지 않습니다."));
+
 
 
         // Post post = createPostRequest.toEntity(author, menu, product, project);
@@ -66,15 +76,13 @@ public class PostService {
                 throw new IllegalArgumentException("Invalid PostType: " + requestType);
             }
         }
-        if (author.getPosition() == Position.INDIVIDUAL) {
-            Post post = createPostRequest.toEntity(author, menu, product, project, postType);
-            postRepository.save(post);
 
-            return post.toPostInfoDTO();
-        } else {
-            //기업인경우
-            throw new AuthorityException("포스트 생성 권한이 없습니다.");
-        }
+        //TODO 해당 프로젝트사람들이면
+        Post post = createPostRequest.toEntity(author, menu, product, project, postType);
+        postRepository.save(post);
+
+        return toPostInfoDTO(post);
+
     }
 
 
@@ -82,9 +90,16 @@ public class PostService {
     Delete
      */
     @Transactional
-    public void deletePost(Long id) {
+    public String deletePost(Long id,Long currentUserId) {
         Post post = postRepository.findById(id).orElseThrow(NotFoundIdException::new);
-        postRepository.delete(post);
+
+        if(post.getAuthor().getId().equals(currentUserId)){
+            postRepository.delete(post);
+            return "delete 완료";
+        }
+        else{
+            throw new AuthorityException("작성자와 일치하지 않아 삭제 권한이 없습니다.");
+        }
     }
 
     /*
@@ -96,7 +111,7 @@ public class PostService {
         Post post = postRepository.findById(id).orElseThrow(NotFoundTaskByIdException::new);
         if(post.getAuthor().getId().equals(currentUserId)){
             post.updatePostTitle(updatePostTitleRequest.getUpdateTitle());
-            return post.toPostInfoDTO();
+            return toPostInfoDTO(post);
         }
         else{
             throw new AuthorityException("작성자와 일치하지 않아 수정 권한이 없습니다.");
@@ -109,7 +124,7 @@ public class PostService {
         Post post = postRepository.findById(id).orElseThrow(NotFoundTaskByIdException::new);
         if(post.getAuthor().getId().equals(currentUserId)){
             post.updatePostContent(updatePostContentRequest.getUpdateContent());
-            return post.toPostInfoDTO();
+            return toPostInfoDTO(post);
         }
         else{
             throw new AuthorityException("작성자와 일치하지 않아 수정 권한이 없습니다.");
@@ -136,7 +151,7 @@ public class PostService {
                 }
             }
             post.updatePostType(updatepostType);
-            return post.toPostInfoDTO();
+            return toPostInfoDTO(post);
         }
         else{
             throw new AuthorityException("작성자와 일치하지 않아 수정 권한이 없습니다.");
@@ -151,7 +166,7 @@ public class PostService {
                 .orElseThrow(() -> new NotFoundIdException("해당 메뉴는 존재하지 않습니다."));
         if(post.getAuthor().getId().equals(currentUserId)){
             post.updatePostMenu(menu);
-            return post.toPostInfoDTO();
+            return toPostInfoDTO(post);
         }
         else{
             throw new AuthorityException("작성자와 일치하지 않아 수정 권한이 없습니다.");
@@ -165,7 +180,7 @@ public class PostService {
     public PostInfoDTO updateProductName(Long id, UpdatePostProductRequest updatePostProductRequest, Long currentUserId) {
         Post post = postRepository.findById(id).orElseThrow(NotFoundTaskByIdException::new);
         post.updatePostProductName(updatePostProductRequest.getUpdateProductName());
-        return post.toPostInfoDTO();
+        return toPostInfoDTO(post);
 
 
     }
@@ -175,7 +190,7 @@ public class PostService {
     public PostInfoDTO updateVersion(Long id, UpdatePostVersionRequest updatePostVersionRequest, Long currentUserId) {
         Post post = postRepository.findById(id).orElseThrow(NotFoundTaskByIdException::new);
         post.updatePostVersion(updatePostVersionRequest.getUpdateVersion());
-        return post.toPostInfoDTO();
+        return toPostInfoDTO(post);
 
     }
 
@@ -186,22 +201,43 @@ public class PostService {
     @Transactional(readOnly = true)
     public PostInfoDTO findById(Long Id){
         Post post=postRepository.findById(Id).orElseThrow(NotFoundIdException::new);;
-        return post.toPostInfoDTO();
+        return toPostInfoDTO(post);
     }
     @Transactional(readOnly = true)
     public List<PostInfoDTO> findPostInfoByMenuId(Long menuId){
         List<Post> postList=postRepository.findByMenuId(menuId);
         List<PostInfoDTO> postInfoDTOs=new ArrayList<>();
         for(Post post:postList){
-            PostInfoDTO postInfoDTO=post.toPostInfoDTO();
+            PostInfoDTO postInfoDTO=toPostInfoDTO(post);
             postInfoDTOs.add(postInfoDTO);
         }
         return postInfoDTOs;
 
     }
+    //이건 안쓰는거->메뉴에서 씀
     public List<Post> findPostsByMenuId(Long menuId) {
         List<Post> postList = postRepository.findByMenuId(menuId);
         return postList;
+    }
+
+    public PostInfoDTO toPostInfoDTO(Post post) {
+        int likeCount = postLikeRepository.countByPostId(post.getId());
+        int commentCount = commentRepository.countByPostId(post.getId());
+
+        return PostInfoDTO.builder()
+                .id(post.getId())
+                .title(post.getTitle())
+                .authorInfoDTO(post.getAuthor().powerMemberInfoDTO())
+                .menuId(post.getMenu().getId())
+                .menuName(post.getMenu().getMenuName())
+                .productName(post.getProductName())
+                .projectName(post.getVersion())
+                .postType(post.getPostType())
+                .content(post.getContent())
+                .createTime(post.getCreateTime())
+                .likeCount(likeCount)
+                .commentCount(commentCount)
+                .build();
     }
